# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fraud Detection & Alert System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List fraud alerts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, high, critical]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved, dismissed]
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlerts"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAlertBody"
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  # ── Cases ─────────────────────────────────────────────────────────────────
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List fraud investigation cases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, pending_sar, closed, escalated]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: assigneeId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Case list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCases"

    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new fraud investigation case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get case details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a fraud case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}/notes:
    post:
      operationId: addCaseNote
      tags: [cases]
      summary: Add an investigation note to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCaseNoteBody"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"

  # ── Rules ─────────────────────────────────────────────────────────────────
...
  /analytics/model-performance:
    get:
      operationId: getModelPerformance
      tags: [analytics]
      summary: Get ML model performance metrics
      parameters:
        - name: modelName
          in: query
          schema:
            type: string
            enum: [isolation_forest, autoencoder, velocity_model]
            default: isolation_forest
      responses:
        "200":
          description: Model performance metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPerformance"
  /analytics/fraud-trends:
    get:
      operationId: getFraudTrends
      tags: [analytics]
      summary: Get fraud trend analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Fraud trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudTrends"
...
  /analytics/top-risk-users:
    get:
      operationId: getTopRiskUsers
      tags: [analytics]
      summary: Get top high-risk users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top risk users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRiskSummary"
...
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
          nullable: true
        userId:
          type: string
          nullable: true
        type:
          type: string
          enum: [velocity_breach, geo_anomaly, device_mismatch, amount_threshold, pattern_match, ml_anomaly, blocked_merchant]
        severity:
          type: string
          enum: [info, warning, high, critical]
        status:
          type: string
          enum: [open, acknowledged, resolved, dismissed]
        title:
          type: string
        description:
          type: string
        ruleId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, type, severity, status, title, description, createdAt]
    PaginatedAlerts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    ResolveAlertBody:
      type: object
      properties:
        resolution:
          type: string
        resolvedBy:
          type: string
      required: [resolution, resolvedBy]

    # ── Case ───────────────────────────────────────────────────────────────
    Case:
      type: object
      properties:
        id:
          type: string
        caseNumber:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, pending_sar, closed, escalated]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        fraudType:
          type: string
        estimatedLoss:
          type: number
        confirmedLoss:
          type: number
          nullable: true
        assigneeId:
          type: string
          nullable: true
        assigneeName:
          type: string
          nullable: true
        transactionIds:
          type: array
          items:
            type: string
        alertIds:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/CaseNote"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        sarId:
          type: string
          nullable: true
      required: [id, caseNumber, title, status, priority, fraudType, estimatedLoss, transactionIds, alertIds, notes, createdAt, updatedAt]
    PaginatedCases:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Case"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        fraudType:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        transactionIds:
          type: array
          items:
            type: string
        alertIds:
          type: array
          items:
            type: string
        assigneeId:
          type: string
          nullable: true
        estimatedLoss:
          type: number
      required: [title, description, fraudType, priority, estimatedLoss]
    UpdateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, pending_sar, closed, escalated]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: string
          nullable: true
        confirmedLoss:
          type: number
          nullable: true
    CaseNote:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, caseId, authorId, authorName, content, createdAt]
    AddCaseNoteBody:
      type: object
      properties:
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
      required: [authorId, authorName, content]

    # ── Rule ───────────────────────────────────────────────────────────────
...
    UserRiskSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        riskScore:
          type: number
        riskTier:
          type: string
        fraudulentTransactions:
          type: integer
        totalSpend:
          type: number
      required: [id, name, email, riskScore, riskTier, fraudulentTransactions, totalSpend]

    # ── Merchant ───────────────────────────────────────────────────────────
...
    TimeSeriesPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        total:
          type: integer
        fraudulent:
          type: integer
        blocked:
          type: integer
        amount:
          type: number
      required: [timestamp, total, fraudulent, blocked, amount]
...
    ModelPerformance:
      type: object
      properties:
        modelName:
          type: string
        accuracy:
          type: number
        precision:
          type: number
        recall:
          type: number
        f1Score:
          type: number
        auc:
          type: number
        falsePositiveRate:
          type: number
        falseNegativeRate:
          type: number
        avgLatencyMs:
          type: number
        totalScored:
          type: integer
        lastTrainedAt:
          type: string
          format: date-time
        confusionMatrix:
          type: object
          properties:
            tp:
              type: integer
            fp:
              type: integer
            tn:
              type: integer
            fn:
              type: integer
          required: [tp, fp, tn, fn]
        scoreDistribution:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
              count:
                type: integer
            required: [bucket, count]
      required: [modelName, accuracy, precision, recall, f1Score, auc, falsePositiveRate, falseNegativeRate, avgLatencyMs, totalScored, lastTrainedAt, confusionMatrix, scoreDistribution]
    FraudTrends:
      type: object
      properties:
        period:
          type: string
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        byFraudType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
              amount:
                type: number
            required: [type, count, amount]
        byCountry:
          type: array
          items:
            type: object
            properties:
              country:
                type: string
              count:
                type: integer
              amount:
                type: number
            required: [country, count, amount]
        trendChange:
          type: number
          description: Percentage change vs previous period
      required: [period, byDay, byFraudType, byCountry, trendChange]
...